Change default idle_time to be 80% of timeout
So that we avoid the issue where the worker would be late in notifying
the master that it is alive.

This is because idle_time is the interval between alive notifications.
If idle_time is greater than timeout then the worker won't have enough
time to notify the master that it is alive.

Change-Id: Ia63462b688f052faa5cc5c83ef25ace2e9b0db1c