Fix: script commands sends to reexec'ed master
The old master's pid file should not be touched until the reexec'ed
master is ready. Then it replaces the old master by killing it then
renaming its pid file as the old master's pid file.

This fixes the problems where the script commands are instead passed to
the reexec master, which I think, is not what most will think as the
intended behaviour.

Change-Id: I3373804dfecfa92226ee14cfb783a2df1532a5ca